Now the log-level is parsed from the commandline and passed down to the ones who care.
It's a trivial implementation, and what I'd be missing is a way to have progress logging independently
form the logging verbosity. However, I believe in most cases, people won't bother anyway.
Even I don't see a real application for this feature ... .
Considering it added some changes which aren't so nice, maybe it shouldn't be there in the first place ?