Fixed GUI for file exclude filters
Everything so far seems to work as expected, finally !
Next up would be the source/destination selection.

This would require a new piece of Rest interface to query folder contents,
some special handling for windows (as it doesn't have a single root),
and of course lots of GUI fancyness in html/java script.

Finally it should be a directive to allow reusing it more easily.
The directive needs to take configuration to allow specifying where
to put the information (using ng-model, I suppose)